Fix content legibility of system bars
For the purpose of maintaining content legibility of system bars
on Theme.Splash, add color on statusBar/navigationBar on each
API version.

* API 21, 22
The statusBarColor/navigationBarColor is black.

* API 23, 24, 25, 26
The statusBarColor is transparent, and windowLightStatusBar=true.
The navigationBarColor is black.

* API 27, 28, 29, 30, 31
The statusBarColor is transparent, and windowLightStatusBar=true.
The navigationBarColor is transparent, and
windowLightNavigationBar=true